refactor: ! Move status menu code to StatusMenu
