+++ src/TaskLineRenderer.ts
@@ -1,9 +1,8 @@
 import type { Moment } from 'moment';
-import { Component, MarkdownRenderer, MenuItem } from 'obsidian';
+import { Component, MarkdownRenderer } from 'obsidian';
 import { GlobalFilter } from './Config/GlobalFilter';
 import { TASK_FORMATS, getSettings } from './Config/Settings';
 import { replaceTaskWithTasks } from './File';
-import { StatusRegistry } from './StatusRegistry';
 import type { Task } from './Task';
 import * as taskModule from './Task';
 import { TaskFieldRenderer } from './TaskFieldRenderer';
@@ -126,34 +125,7 @@ export class TaskLineRenderer {
         });
 
         checkbox.addEventListener('contextmenu', async (ev: MouseEvent) => {
-            const menu = new StatusMenu();
-            // TODO Add a tooltip, so it's more obvious that right-click is available
-            const commonTitle = 'Change status to: ';
-
-            const getMenuItemCallback = (item: MenuItem, statusName: string, newStatusSymbol: string) => {
-                item.setTitle(`${commonTitle}  ${statusName}`).onClick(() => {
-                    // TODO Don't make a change if the status is already set to this value.
-                    const status = StatusRegistry.getInstance().bySymbol(newStatusSymbol);
-                    const newTask = task.handleStatusChangeFromContextMenuWithRecurrenceInUsersOrder(status);
-                    replaceTaskWithTasks({
-                        originalTask: task,
-                        newTasks: newTask,
-                    });
-                });
-            };
-
-            // TODO Don't use StatusSettings: they are unchecked and can have duplicates and errors.
-            //      Use StatusRegistry.getInstance() instead.
-            // TODO Put a checkmark against the current status symbol.
-            // TODO Show the status symbol
-            // TODO Maybe group by status type?
-            const { statusSettings } = getSettings();
-            for (const status of statusSettings.coreStatuses) {
-                menu.addItem((item) => getMenuItemCallback(item, status.name, status.symbol));
-            }
-            for (const status of statusSettings.customStatuses) {
-                menu.addItem((item) => getMenuItemCallback(item, status.name, status.symbol));
-            }
+            const menu = new StatusMenu(task);
 
             menu.showAtPosition({ x: ev.clientX, y: ev.clientY });
         });
