+++ tests/__mocks__/obsidian.ts
@@ -1,3 +1,13 @@
 export {};
 
-export class Menu {}
+export class MenuItem {}
+
+export class Menu {
+    /**
+     * Adds a menu item. Only works when menu is not shown yet.
+     * @public
+     */
+    addItem(_cb: (item: MenuItem) => any): this {
+        return this;
+    }
+}
